feat: add db configurations
