--- db/database.py
@@ -1,1 +1,0 @@
-# from sqlmodel import create_engine
